feat(refactoring): Code refactoring completed 100% (for now :p)
